pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "baby-macropad"
version = "0.3.0"
description = "Stream Dock M18 macropad controller for Baby Basics baby tracking"
readme = "README.md"
license = "MIT"
requires-python = ">=3.11"
dependencies = [
    "httpx>=0.27",
    "Pillow>=10.0",
    "pydantic>=2.0",
    "pydantic-settings>=2.0",
    "PyYAML>=6.0",
    "qrcode>=7.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.24",
    "respx>=0.22",
    "ruff>=0.8",
]

[project.scripts]
baby-macropad = "baby_macropad.main:main"

[tool.hatch.build.targets.wheel]
packages = ["src/baby_macropad"]

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"

[tool.ruff]
target-version = "py311"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP"]
